requirements.txt
dj-database-url==2.0.0
django==3.2
#psycopg2-binary==2.9.3
gunicorn==20.1.0
wheel==0.37.0
whitenoise==6.0.0
requests
django-phonenumber-field[phonenumbers]==7.3.0
django-countries==7.5.1
pillow
babel
boto3
django-storages
yfinance==0.2.40
